Upgraded tradingeconomics to v3 framework
