+++ .changeset/tricky-experts-mix.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/tradingeconomics-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
