--- packages/sources/tradingeconomics-test/CHANGELOG.md
@@ -1,52 +1,0 @@
-# @chainlink/tradingeconomics-test-adapter
-
-## 1.0.7
-
-### Patch Changes
-
-- 011aec39e: Bumped framework version
-
-## 1.0.6
-
-### Patch Changes
-
-- 1a00fdfc1: Bumped framework version
-- 715221438: Bumped framework version
-
-## 1.0.5
-
-### Patch Changes
-
-- eff4c4cf5: Bumped framework version
-- 14a549217: Bumped framework version. Replaced PriceEndpoint with ForexPriceEndpoint and AdapterEndpoint with StockEndpoint for relevant endpoints
-
-## 1.0.4
-
-### Patch Changes
-
-- 244e02abf: Refactored file structure
-
-## 1.0.3
-
-### Patch Changes
-
-- 862ed8d89: Removed unused dependencies
-
-## 1.0.2
-
-### Patch Changes
-
-- cc0d39064: Bumped framework version
-- fb1b11b77: Bumped framework version
-
-## 1.0.1
-
-### Patch Changes
-
-- 076976778: Fixed symbol casing
-
-## 1.0.0
-
-### Major Changes
-
-- 3469ed6c4: Add tradingeconomics v3 EA
