--- packages/sources/tradingeconomics-test/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@chainlink/tradingeconomics-test-adapter",
-  "version": "1.0.7",
-  "description": "Chainlink tradingeconomics-test adapter.",
-  "keywords": [
-    "Chainlink",
-    "LINK",
-    "blockchain",
-    "oracle",
-    "tradingeconomics-test"
-  ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "repository": {
-    "url": "https://github.com/smartcontractkit/external-adapters-js",
-    "type": "git"
-  },
-  "license": "MIT",
-  "scripts": {
-    "clean": "rm -rf dist && rm -f tsconfig.tsbuildinfo",
-    "prepack": "yarn build",
-    "build": "tsc -b",
-    "server": "node -e 'require(\"./index.js\").server()'",
-    "server:dist": "node -e 'require(\"./dist/index.js\").server()'",
-    "start": "yarn server:dist"
-  },
-  "dependencies": {
-    "@chainlink/external-adapter-framework": "0.31.0",
-    "tslib": "^2.3.1"
-  },
-  "devDependencies": {
-    "@sinonjs/fake-timers": "9.1.2",
-    "@types/jest": "27.5.2",
-    "@types/node": "16.11.51",
-    "@types/sinonjs__fake-timers": "8.1.2",
-    "nock": "13.2.9",
-    "typescript": "5.0.4"
-  }
-}
