--- packages/sources/tradingeconomics-test/src/config/includes.json
@@ -1,123 +1,0 @@
-[
-  {
-    "from": "CAD",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "CAD",
-        "inverse": true
-      }
-    ]
-  },
-  {
-    "from": "CHF",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "CHF",
-        "inverse": true
-      }
-    ]
-  },
-  {
-    "from": "CNY",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "CNY",
-        "inverse": true
-      }
-    ]
-  },
-  {
-    "from": "JPY",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "JPY",
-        "inverse": true
-      }
-    ]
-  },
-  {
-    "from": "KRW",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "KRW",
-        "inverse": true
-      }
-    ]
-  },
-  {
-    "from": "MXN",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "MXN",
-        "inverse": true
-      }
-    ]
-  },
-  {
-    "from": "PLN",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "PLN",
-        "inverse": true
-      }
-    ]
-  },
-  {
-    "from": "SEK",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "SEK",
-        "inverse": true
-      }
-    ]
-  },
-  {
-    "from": "SGD",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "SGD",
-        "inverse": true
-      }
-    ]
-  },
-  {
-    "from": "TRY",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "TRY",
-        "inverse": true
-      }
-    ]
-  },
-  {
-    "from": "ZAR",
-    "to": "USD",
-    "includes": [
-      {
-        "from": "USD",
-        "to": "ZAR",
-        "inverse": true
-      }
-    ]
-  }
-]
