--- packages/sources/tradingeconomics-test/src/endpoint/index.ts
@@ -1,2 +1,0 @@
-export { endpoint as price } from './price'
-export { endpoint as stock } from './stock'
