--- packages/sources/tradingeconomics-test/test-payload.json
@@ -1,6 +1,0 @@
-{
- "requests": [{
-  "from": "CAD",
-  "to": "USD"
- }]
-}
