--- packages/sources/tradingeconomics/schemas/env.json
@@ -1,30 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/tradingeconomics-adapter.json",
-  "title": "Chainlink External Adapter for Tradingeconomics",
-  "description": "This adapter uses the Tradingeconomics WS stream",
-  "required": ["API_CLIENT_KEY", "API_CLIENT_SECRET"],
-  "type": "object",
-  "properties": {
-    "API_URL": {
-      "default": "wss://stream.tradingeconomics.com/ or https://api.tradingeconomics.com/markets",
-      "type": "string",
-      "description": "The URL of the WS endpoint",
-      "format": "uri"
-    },
-    "API_CLIENT_KEY": {
-      "type": "string",
-      "required": true,
-      "description": "Your API client key"
-    },
-    "API_CLIENT_SECRET": {
-      "type": "string",
-      "required": true,
-      "description": "Your API client secret"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
