--- packages/sources/tradingeconomics/src/config/symbols.json
@@ -1,6 +1,0 @@
-{
-  "tradingeconomics": {
-    "FTSE": "UKX:IND",
-    "N225": "NKY:IND"
-  }
-}
