+++ packages/sources/tradingeconomics/src/endpoint/index.ts
@@ -1,6 +1,2 @@
-import type { TInputParameters as PriceInputParameters } from './price'
-
-export type TInputParameters = PriceInputParameters
-
-export * as price from './price'
-export * as price_ws from './price-ws'
+export { endpoint as price } from './price'
+export { endpoint as stock } from './stock'
