+++ packages/sources/tradingeconomics/src/endpoint/stock.ts
