+++ packages/sources/tradingeconomics/src/transport/stock-http.ts
