+++ packages/tsconfig.json
@@ -531,9 +531,6 @@
       "path": "./sources/tradingeconomics"
     },
     {
-      "path": "./sources/tradingeconomics-test"
-    },
-    {
       "path": "./sources/trueusd"
     },
     {
