+++ packages/tsconfig.test.json
@@ -531,9 +531,6 @@
       "path": "./sources/tradingeconomics/tsconfig.test.json"
     },
     {
-      "path": "./sources/tradingeconomics-test/tsconfig.test.json"
-    },
-    {
       "path": "./sources/trueusd/tsconfig.test.json"
     },
     {
